Run a report

Run a report. This will create a new directory in drafts/<reportname>, copy your declared resources there, run your script and check that all expected artefacts were created.

orderly_run(
  name,
  parameters = NULL,
  envir = NULL,
  logging_console = NULL,
  logging_threshold = NULL,
  root = NULL,
  locate = TRUE
)

Arguments

name

Name of the report to run

parameters

Parameters passed to the report. A named list of parameters declared in the orderly.yml. Each parameter must be a scalar character, numeric, integer or logical.

envir

The environment that will be used to evaluate the report script; by default we use the global environment, which may not always be what is wanted.

logging_console

Optional logical, passed through to outpack::outpack_packet_start to control printing logs to the console, overriding any default configuration set at the root.

logging_threshold

Optional logging threshold, passed through to outpack::outpack_packet_start to control the amount of detail in logs printed during running, overriding any default configuration set at the root. If given, must be one of info, debug or trace (in increasing order of verbosity). Practically this has no effect at present as we've not added any fine-grained logging.

root

The path to an orderly root directory, or NULL (the default) to search for one from the current working directory if locate is TRUE.

locate

Logical, indicating if the configuration should be searched for. If TRUE and config is not given, then orderly looks in the working directory and up through its parents until it finds an .outpack directory

Value

The id of the created report (a string)